testdocker/requirements.txt
apache-airflow
click>=8.0
cached_property;python_version<"3.8"
jinja2>=2.11.3
keyring==10.1
PyGithub
jsonpath_ng
jsonschema
pendulum
pyyaml
packaging
rich
rich_click
semver
tabulate
requests
gusty==0.3.3 